+++ Qitom/python/qDebugStream.cpp
@@ -65,7 +65,7 @@ std::streamsize QDebugStream::xsputn(const char *p, std::streamsize n)
     int pos = 0;
     while (pos != std::string::npos)
     {
-        pos = m_string.find('\n');
+        pos = (int)m_string.find('\n');
         if (pos != std::string::npos)
         {
             std::string tmp(m_string.begin(), m_string.begin() + pos);
